feat: Enhance leave type configuration with fixed ranges
Introduces a new `DateRange` type and updates `LeaveTypeConfig` to support multiple fixed date ranges per leave type. This allows for more flexible configuration of recurring or predefined leave periods.

The `fixedRange` property in `LeaveTypeConfig` has been renamed to `fixedRanges` and is now an array of `DateRange` objects. Backward compatibility for the old single `fixedRange` object is maintained in the store.

A new state variable `selectedRangeIndex` has been added to `RequestFormModal` to manage the selection of multiple fixed ranges.